refactor: rename time class
