+++ src/test/java/roomescape/reservationTime/dto/ReservationTimeRequestTest.java
@@ -1,16 +1,16 @@
-package roomescape.time.dto;
+package roomescape.reservationTime.dto;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.LocalTime;
 import org.junit.jupiter.api.Test;
-import roomescape.time.ReservationTime;
+import roomescape.reservationTime.ReservationTime;
 
-class TimeRequestTest {
+class ReservationTimeRequestTest {
 
     @Test
     void toDomain_변환() {
-        TimeRequest request = new TimeRequest(LocalTime.of(10, 0));
+        ReservationTimeRequest request = new ReservationTimeRequest(LocalTime.of(10, 0));
 
         ReservationTime time = request.toDomain(1L);
 
@@ -20,7 +20,7 @@ class TimeRequestTest {
 
     @Test
     void toDomain_id가_주입된다() {
-        TimeRequest request = new TimeRequest(LocalTime.of(10, 0));
+        ReservationTimeRequest request = new ReservationTimeRequest(LocalTime.of(10, 0));
 
         ReservationTime first = request.toDomain(1L);
         ReservationTime second = request.toDomain(2L);
