+++ src/test/java/roomescape/reservationTime/dto/ReservationTimeResponseTest.java
@@ -1,18 +1,18 @@
-package roomescape.time.dto;
+package roomescape.reservationTime.dto;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.LocalTime;
 import org.junit.jupiter.api.Test;
-import roomescape.time.ReservationTime;
+import roomescape.reservationTime.ReservationTime;
 
-class TimeResponseTest {
+class ReservationTimeResponseTest {
 
     @Test
     void from_변환() {
         ReservationTime time = new ReservationTime(1L, LocalTime.of(10, 0));
 
-        TimeResponse response = TimeResponse.from(time);
+        ReservationTimeResponse response = ReservationTimeResponse.from(time);
 
         assertThat(response.id()).isEqualTo(1L);
         assertThat(response.startAt()).isEqualTo(LocalTime.of(10, 0));
@@ -22,7 +22,7 @@ class TimeResponseTest {
     void from_도메인_필드가_모두_응답에_포함된다() {
         ReservationTime time = new ReservationTime(42L, LocalTime.of(15, 40));
 
-        TimeResponse response = TimeResponse.from(time);
+        ReservationTimeResponse response = ReservationTimeResponse.from(time);
 
         assertThat(response.id()).isEqualTo(time.getId());
         assertThat(response.startAt()).isEqualTo(time.getStartAt());
